Fix code complexity
